Move signing files constants in trezor tests to separate files
